MemfaultScope moved to internal package
