+++ lib/observability/src/main/java/no/nordicsemi/memfault/observability/internal/MemfaultScope.kt
@@ -29,7 +29,7 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package no.nordicsemi.memfault.observability
+package no.nordicsemi.memfault.observability.internal
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
